+++ Assets/Scripts/XRCamera.cs
@@ -1,6 +1,5 @@
 using DT.UniStart;
 using UnityEngine;
-using UnityRawInput;
 
 public class XRCamera : CBC {
   void Start() {
@@ -9,20 +8,6 @@ public class XRCamera : CBC {
     XRSDK.SetPanelLuminance(Config.instance.PanelLuminance);
 
     var eb = this.Get<IEventBus>();
-    RawInput.WorkInBackground = true;
-    RawInput.Start();
-    if (Config.instance.ResetViewHotKey.Enabled) {
-      RawInput.OnKeyDown += (key) => {
-        if ((uint)key == Config.instance.ResetViewHotKey.Key) {
-          if ((Config.instance.ResetViewHotKey.Modifier & 0x01) != 0 && !RawInput.IsKeyDown(RawKey.LeftButtonAlt)) return;
-          if ((Config.instance.ResetViewHotKey.Modifier & 0x02) != 0 && !RawInput.IsKeyDown(RawKey.LeftControl)) return;
-          if ((Config.instance.ResetViewHotKey.Modifier & 0x04) != 0 && !RawInput.IsKeyDown(RawKey.LeftShift)) return;
-          if ((Config.instance.ResetViewHotKey.Modifier & 0x08) != 0 && !RawInput.IsKeyDown(RawKey.LeftWindows)) return;
-          XRSDK.Reset();
-          eb.Invoke("tip", "Reset View");
-        }
-      };
-    }
 
     this.onUpdate.AddListener(() => {
       // update the camera position by reading the sensor data
@@ -41,10 +26,6 @@ public class XRCamera : CBC {
         XRSDK.Reset();
         eb.Invoke("tip", "Reset View");
       }
-    });
-
-    this.onApplicationQuit.AddListener(() => {
-      RawInput.Stop();
     });
   }
 }